+++ miracle/web/settings.py
@@ -6,7 +6,7 @@ loglevel = 'warning'
 max_requests = 1000000
 max_requests_jitter = max_requests // 10
 timeout = 60
-worker_class = 'contextgraph.web.worker.GeventWorker'
+worker_class = 'miracle.web.worker.GeventWorker'
 worker_connections = 20
 workers = multiprocessing.cpu_count()
 
@@ -16,7 +16,7 @@ def post_worker_init(worker):  # pragma: no cover
 
 
 def worker_exit(server, worker):  # pragma: no cover
-    from contextgraph.web.app import worker_exit
+    from miracle.web.app import worker_exit
     worker_exit(server, worker)
 
 # cleanup