+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.0.2 (unreleased)
 ==================
 
+- Rename to miracle.
+
 
 1.0.1 (2016-06-06)
 ==================